railway

+++ src/main/java/com/project/railway/repository/Route_Repository.java
@@ -4,7 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.project.railway.dto.Route;
 
-public interface Route_Repository extends JpaRepository<Route, Integer> {
+public interface Route_Repository extends JpaRepository<Route, Long> {
 
-	
 }
